Fix comments and make ready to merge
